backend/pyproject.toml
[project]
name = "tailord-backend"
version = "0.1.0"
description = "FastAPI backend for Tailord"
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "playwright",
  "openai",
  "slowapi",
  "python-dotenv",
  "pydantic_settings",
  "beautifulsoup4>=4.14.3",
  "readability-lxml>=0.8.4.1",
  "markdownify>=1.2.2",
  "requests>=2.32.5",
]

[tool.uv]
dev-dependencies = [
  "pytest",
]
